Moving new SHA1 logic into a common location and updating references to use it.
