+++ src/LibZ.Bootstrap.35/LibZ.Bootstrap.35.csproj
@@ -42,6 +42,12 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\LibZ.Utils\LibZ.Utils.csproj">
+      <Project>{9977663f-ab99-4e4e-9677-92a7539e01d1}</Project>
+      <Name>LibZ.Utils</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>copy /y /d $(TargetPath) $(ProjectDir)..\LibZ.Tool.InjectIL\Resources\LibZ.Bootstrap.35.dll</PostBuildEvent>
